DX-11444: Prevent optimization of non-deterministic functions
RexSimplify and RexProgram optimize expressions like A() = A() to true,
but if A() is a non-deterministic expression, this optimization is invalid.
Change RexSimplify and RexProgram to check if expression is deterministic
or not, in order to decide about the optimization.

Note that linq4j has the same exact issue, as exhibited by the test case
added to misc.iq. That said, the amount of change required to fix that last
issue is not small, and considering how Dremio do not use that code, test
is temporarily commented until a solution is figured out with upstream.

Change-Id: I8a30e9d35c6439ad767547f02d3646146910b928