DX-11490: Fix wrong constant reduction over windows function
Calcite was incorrectly inferring a IS NULL predicate would cause
a reference over a window function to return null, but a window function
type is never nullable.

Fixing in RexUtil to check nullability of the expression before reduction
to not add the wrong constant reduction if type is not nullable.

Change-Id: Ib0a61803d2d56866c13279b648e560885346add7