DX-11741: Fix infinite loop caused by PushProjector
Although PushProject preserves OVER() expression and pushes it
into children projects, it doesn't replace the expression by a
reference in the original top project, causing Hep planner to go
into a loop until a StackOverflowError is thrown.

Fix the issue by changing RefAndExprConverter visitor to check
if the RexOver call is in the list of expressions to preserve, like
any other RexCall node.

Change-Id: Iacefd854011d0677eeee7a7806e33a2520f2eb9b