[CALCITE-2339] Add utility to convert datetime_plus to timestampadd (James Duong)
Canonicalize the representation of adding a single time unit
to a datetime type under the datetime_plus operator. Add a RexUtil
method for converting back to timestampadd.

Add a hook that lets a SqlDialect request datetime_plus to be
exposed as timestampadd.

This work duplicates some of the work in [CALCITE-2188],
however it allows for this conversion to happen outside of
SqlDialect. It also handles the case where an interval is not
just a literal, but a multiplication of an interval literal by an
integer (the TimestampAddConvertlet does this when
converting TIMESTAMPADD to DATETIME_PLUS initially)

We will preserve WEEK and QUARTER units passed to
TIMESTAMPADD.

Fix unparsing of WEEK and QUARTER intervals produced by the
TimestampAddConvertlet when using Db2Dialect and SqlDialect

Change-Id: I0d80d0dff45304ba71290ad8c0a9f5502cd6ec87