[CALCITE-2471] Shorten AbstractRelNode#computeDigest() digest (Laurent Goujon)
Digest/description created by AbstractRelNode#computeDigest includes
the description of each of its inputs, causing the final string to
be excessively long (e.g. the top level node of a tree containing
the string representation of the whole tree).

Change the function to not use the input's description but instead
its original description which is <typeName>#<id>.

fixes #795

Change-Id: Id09b741a9db8374f127c0c51c9a2766cadfcf2c5