[CALCITE-2437] FilterMultiJoinMerge doesn't combine filters
The FilterMultiJoinMergeRule class is intended to combine a filter on
top of a multi join into a single multijoin. But instead of combining
the post join filter with the filter's condition, it simply replaces
it.

Fixing the issue by composing a conjuction of the two filters and
adding it to the new multijoin.

Change-Id: Ib741f9b9cf490259c9dd104308802f6f85975e4f